refactor: remove the output port directory for router
